Modified the merging code to pass when the file exists

+++ merge_time_CDO_CMIP6.sh
@@ -39,28 +39,34 @@ for var in $var_list; do
         echo "OUTPUT : "$file_output
         echo ""
 
-        export num_model_files=$(ls ${path_input} | grep ${input_prefix} | wc -l)
-        echo "Number of variable("$var") files in model("$model"): " $num_model_files
+        # If output file exists, go to next model & variable or remove it and retry
+        if [ -f "$file_output" ] ; then 
+            echo "OUTPUT file exists already ! Remove it and retry"
+            continue
+        else
+            export num_model_files=$(ls ${path_input} | grep ${input_prefix} | wc -l)
+            echo "Number of variable("$var") files in model("$model"): " $num_model_files
 
-        if [ $num_model_files -eq 1 ]; then 
-            # Copy the CMIP6 file, if there is only 1 file 
-            cp $file_input"_"$year_str"01-"$year_end"12.nc" $file_output
-        fi
+            if [ $num_model_files -eq 1 ]; then 
+                # Copy the CMIP6 file, if there is only 1 file 
+                cp $file_input"_"$year_str"01-"$year_end"12.nc" $file_output
+            fi
 
-        if [ $num_model_files -gt 1 ] ; then 
-            if [ ${scenario} = "ssp126" ] || [ ${scenario} = "ssp245" ] || \
-               [ ${scenario} = "ssp585" ]; then 
-                # Exception for IPSL / MRI model, which extends to 2300 year in ScenarioMIP
-                if [ ${model} = "IPSL-CM6A-LR" ] || [ ${model} = "MPI-ESM2-0" ]; then 
-                    cp $file_input"_"$year_str"01-"$year_end"12.nc" $file_output
+            if [ $num_model_files -gt 1 ] ; then 
+                if [ ${scenario} = "ssp126" ] || [ ${scenario} = "ssp245" ] || \
+                   [ ${scenario} = "ssp585" ]; then 
+                    # Exception for IPSL / MRI model, which extends to 2300 year in ScenarioMIP
+                    if [ ${model} = "IPSL-CM6A-LR" ] || [ ${model} = "MPI-ESM2-0" ]; then 
+                        cp $file_input"_"$year_str"01-"$year_end"12.nc" $file_output
+                    else
+                    # Excute the time merging by using CDO commands
+                        cdo mergetime $file_input"*" $file_output
+                    fi
                 else
                 # Excute the time merging by using CDO commands
                     cdo mergetime $file_input"*" $file_output
-                fi
-            else
-            # Excute the time merging by using CDO commands
-                cdo mergetime $file_input"*" $file_output
-            fi 
+                fi 
+            fi
         fi
 
         echo "MERGING COMPLETED "
